1950: fixed menu style issues + normalize + PHPinc
The menu was misbehaving, so I made it behave better.
Still not sure why we have it like that, rather than simple jQuery.

Added normalize, because if we are going to have a reset, might as well

Added support for includes, and because it's an index.php
it should load whenever someone comes to the lecture10 folder.